feat(kanban): initial kanban commit
This commit introduces an initial prototype for the kanban board
functionality. The code itself is very rough. It grabs all the yaml
front matter for the notes, figures out what projects are present and
has a two screen tui application where the first displays the projects
and the second has columns like a Kanban board. Currently the second
screen is filled with dummy data.

Refs: