feat(kanban tui): change the keybind to go to the dashboard
This commit changes the keybind to return to the top dashboard from
escape to control+d. This was done because the escape key was requiring
double presses to first escape the current buffer then to activate the
custom keybind. Control+d is more directly and only has to be performed
once.

Refs: